+++ duality/decorators/__init__.py
@@ -1,11 +1,11 @@
 # imports relevant contents.
-from duality.plugins.metaclass import Meta
+from vandal.plugins.metaclass import Meta
 from duality.decorators.record import Record
 from duality.decorators.track import Track
 
 # all relevant contents.
 __all__ = [
-    Meta,
-    Record,
-    Track,
+    'Meta',
+    'Record',
+    'Track',
 ]