+++ duality/decorators/track.py
@@ -1,7 +1,7 @@
 # makes multiple instances of the object available.
-from duality.plugins.metaclass import Meta
+from vandal.plugins.metaclass import Meta
 
-from duality.plugins.types import (
+from vandal.plugins.types import (
     VandalType,
     IntegerType,
     FloatType,